Abstraction of platform-specific details
Those aspects of that would determine the input to functions e.g. time-series plotting are put together in classes e.g. Choice of columns for analysis, handling of date-time columns, plotting parameters, etc. At this point, this is quite rudimentary and will be extended as complexity increases in the package.